requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.0
git-filter-repo==2.38.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-secret-manager==2.20.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.31.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0